+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Added
 
+- added `make_py_pkg` function <2024-10-22, Tue>
+- added `distance2df` function <2024-10-22, Tue>
 - added `euler` plot for venn <2024-07-20, Sat>
 - added `igraph_translator` function <2024-07-11, Thu>